[Security Solution][Endpoint][Response Actions] Additional actions log tests (#141447)
* list handler tests

- tests the logic for branching with status filter
- tests request params formats for strings into string []
- tests logic for index_not_found_exception

* verify that hostnames filter also shows inactive agents

refs elastic/security-team/issues/4721

* verify useGetEndpointsList lists 50 agents at most

Even when metadata list has more than 50 agents the `useGetEndpointsList` hook should list only 50 at most

refs elastic/security-team/issues/4721

* verify useGetEndpointsList shows 10 more in list

When all 50 or more agents are selected in the filter list, the hook should list 10 more agents

refs elastic/security-team/issues/4721

* Add a metadata list test for hostname search

refs elastic/security-team/issues/4721

* Refactor table props into a hook

fixes elastic/security-team/issues/4990

* remove redundant checks

review changes (@paul-tavares)

(cherry picked from commit cbb2e2a4c6d181b87dffafba495cc391ceeb9da7)
