+++ x-pack/plugins/security_solution/public/management/hooks/endpoint/use_get_endpoints_list.test.ts
@@ -11,6 +11,8 @@ import { useGetEndpointsList } from './use_get_endpoints_list';
 import { HOST_METADATA_LIST_ROUTE } from '../../../../common/endpoint/constants';
 import { useQuery as _useQuery } from '@tanstack/react-query';
 import { endpointMetadataHttpMocks } from '../../pages/endpoint_hosts/mocks';
+import { EndpointStatus, HostStatus } from '../../../../common/endpoint/types';
+import { EndpointDocGenerator } from '../../../../common/endpoint/generate_data';
 
 const useQueryMock = _useQuery as jest.Mock;
 
@@ -106,5 +108,128 @@ describe('useGetEndpointsList hook', () => {
         enabled: false,
       })
     );
+  });
+
+  it('should also list inactive agents', async () => {
+    const getApiResponse = apiMocks.responseProvider.metadataList.getMockImplementation();
+
+    // set a few of the agents as inactive/unenrolled
+    apiMocks.responseProvider.metadataList.mockImplementation(() => {
+      if (getApiResponse) {
+        return {
+          ...getApiResponse(),
+          data: getApiResponse().data.map((item, i) => {
+            const isInactiveIndex = [0, 1, 3].includes(i);
+            return {
+              ...item,
+              host_status: isInactiveIndex ? HostStatus.INACTIVE : item.host_status,
+              metadata: {
+                ...item.metadata,
+                host: {
+                  ...item.metadata.host,
+                  hostname: isInactiveIndex
+                    ? `${item.metadata.host.hostname}-inactive`
+                    : item.metadata.host.hostname,
+                },
+                Endpoint: {
+                  ...item.metadata.Endpoint,
+                  status: isInactiveIndex
+                    ? EndpointStatus.unenrolled
+                    : item.metadata.Endpoint.status,
+                },
+              },
+            };
+          }),
+        };
+      }
+      throw new Error('some error');
+    });
+
+    // verify useGetEndpointsList hook returns the same inactive agents
+    const res = await renderReactQueryHook(() => useGetEndpointsList({ searchString: 'inactive' }));
+    expect(
+      res.data?.map((host) => host.name.split('-')[2]).filter((name) => name === 'inactive').length
+    ).toEqual(3);
+  });
+
+  it('should only list 50 agents when more than 50 in the metadata list API', async () => {
+    const getApiResponse = apiMocks.responseProvider.metadataList.getMockImplementation();
+
+    apiMocks.responseProvider.metadataList.mockImplementation(() => {
+      if (getApiResponse) {
+        const generator = new EndpointDocGenerator('seed');
+        const total = 60;
+        const data = Array.from({ length: total }, () => {
+          const endpoint = {
+            metadata: generator.generateHostMetadata(),
+            host_status: HostStatus.UNHEALTHY,
+          };
+
+          generator.updateCommonInfo();
+
+          return endpoint;
+        });
+
+        return {
+          ...getApiResponse(),
+          data,
+          page: 0,
+          // this page size is not used by the hook (it uses the default of 50)
+          // this is only for the test
+          pageSize: 80,
+          total,
+        };
+      }
+      throw new Error('some error');
+    });
+
+    // verify useGetEndpointsList hook returns all 50 agents in the list
+    const res = await renderReactQueryHook(() => useGetEndpointsList({ searchString: '' }));
+    expect(res.data?.length).toEqual(50);
+  });
+
+  it('should only list 10 more agents when 50 or more agents are already selected', async () => {
+    const getApiResponse = apiMocks.responseProvider.metadataList.getMockImplementation();
+
+    apiMocks.responseProvider.metadataList.mockImplementation(() => {
+      if (getApiResponse) {
+        const generator = new EndpointDocGenerator('seed');
+        const total = 61;
+        const data = Array.from({ length: total }, () => {
+          const endpoint = {
+            metadata: generator.generateHostMetadata(),
+            host_status: HostStatus.UNHEALTHY,
+          };
+
+          generator.updateCommonInfo();
+
+          return endpoint;
+        });
+
+        return {
+          ...getApiResponse(),
+          data,
+          page: 0,
+          // since we're mocking that all 50 agents are selected
+          // page size is set to max allowed
+          pageSize: 10000,
+          total,
+        };
+      }
+      throw new Error('some error');
+    });
+
+    // get the first 50 agents to select
+    const agentIdsToSelect = apiMocks.responseProvider
+      .metadataList()
+      .data.map((d) => d.metadata.agent.id)
+      .slice(0, 50);
+
+    // call useGetEndpointsList  with all 50 agents selected
+    const res = await renderReactQueryHook(() =>
+      useGetEndpointsList({ searchString: '', selectedAgentIds: agentIdsToSelect })
+    );
+    // verify useGetEndpointsList hook returns 60 agents
+    expect(res.data?.length).toEqual(60);
   });
 });
